Conflicts:
	AlfrescoApp/View Controllers/Metadata View Controller/MetaDataViewController.m
	AlfrescoApp/View Controllers/Tabbar View Controllers/Task View Controller/TaskViewController.m

+++ AlfrescoApp/Constants/Constants.m
@@ -8,7 +8,7 @@
 
 #import "Constants.h"
 
-NSInteger const kMaxItemsPerListingRetrieve = 25;
+int const kMaxItemsPerListingRetrieve = 25;
 
 NSString * const kLicenseDictionaries = @"thirdPartyLibraries";
